Create specializable codi::addressof function
The function returns address of a variable, and is used in
tapes/chunk.hpp. The motivation for this is that, for example, the
addressof operation is overloaded by boost::proto for vex::symbolic
types and expressions, and so does not return an address of the
variables, but a new expression. With this commit it is possible to
specialize codi::addressof_impl for such cases.