Critical bugfix for index manager.
The index manager produced errors if a statement in the code
does the following:
  a = a;

In the method first the lhs index was freed and afterwards the rhs
index was copied to the lhs. If the adresses of both indices are
the same then after the free of the lhs index, the rhs index is
also zero. This caused the effect that this statement disabled
a as an active variable.