Feature: Helpers for linear system solvers.
Integrates helpers for the external function handling of linear system
solvers. The LinearSystemInterface interface needs to be implemented for
the linear system solver that needs to be handled. This specialization
needs then to be called with:
  codi::solveLinearSystem(specialization, matrix, rhs, sol);

A specialization for Eigen is available. Here, only the solveSystem
method needs to be specialized. See Example 21 for details.

Merge pull request #29 from feature/linearSystemSolvers: Helpers for linear system solvers
Reviewed-by: Johannes Blühdorn <johannes.bluehdorn@scicomp.uni-kl.de>