Changes for clangd compatibility of IDE mode.
This patch removes most of the errors and warnings clangd produces when
CODI_IDE is used. In some places this makes it necessary to not declare
a default argument for templates.

This patch introduces CODI_DEFAULT_TAPE and CODI_DEFAULT_LHS_EXPRESSION
macros for the default declaration of a CoDiPack tape or lhs value.

Merge pull request #33 from feature/ideAutocompleteClangd: Changes for IDE mode compatibility with clangd.
Reviewed-by: Johannes Blühdorn <johannes.bluehdorn@scicomp.uni-kl.de>