Add tape write and read functionality.
Introduces a new tape interface for reading and writing tapes to and
from disk. The written tapes are standalone and can be read in a
different context. See example 25, 26, and 27 for details.

Custom tape readers and writers can be implemented by extending the
TapeWriterInterface and TapeRaderInterface.

Merge pull request #48 from 'feature/tapeWriteAndRead' into develop (squashed)
Reviewed-by: Max Sagebaum <max.sagebaum@scicomp.uni-kl.de>